refactor adherence report generation out of view module.
